Conflicts:
	src/SQLite/cSQLiteOutputManager.cc
	src/postgreSQL/cPorstgreSQLOutputManager.cc

+++ tools/SQLiteDatabaseMerge/SQLiteDatabaseMerge.py
@@ -1,4 +1,34 @@
 #!/usr/bin/env python
+# Copyright (c) 2016, CoRE Research Group, Hamburg University of Applied Sciences
+# All rights reserved.
+# 
+# Redistribution and use in source and binary forms, with or without modification,
+# are permitted provided that the following conditions are met:
+# 
+# 1. Redistributions of source code must retain the above copyright notice, this
+#    list of conditions and the following disclaimer.
+# 
+# 2. Redistributions in binary form must reproduce the above copyright notice,
+#    this list of conditions and the following disclaimer in the documentation
+#    and/or other materials provided with the distribution.
+# 
+# 3. Neither the name of the copyright holder nor the names of its contributors
+#    may be used to endorse or promote products derived from this software without
+#    specific prior written permission.
+# 
+# THIS SOFTWARE IS PROVIDED BY THE COPYRIGHT HOLDERS AND CONTRIBUTORS "AS IS" AND
+# ANY EXPRESS OR IMPLIED WARRANTIES, INCLUDING, BUT NOT LIMITED TO, THE IMPLIED
+# WARRANTIES OF MERCHANTABILITY AND FITNESS FOR A PARTICULAR PURPOSE ARE
+# DISCLAIMED. IN NO EVENT SHALL THE COPYRIGHT HOLDER OR CONTRIBUTORS BE LIABLE FOR
+# ANY DIRECT, INDIRECT, INCIDENTAL, SPECIAL, EXEMPLARY, OR CONSEQUENTIAL DAMAGES
+# (INCLUDING, BUT NOT LIMITED TO, PROCUREMENT OF SUBSTITUTE GOODS OR SERVICES;
+# LOSS OF USE, DATA, OR PROFITS; OR BUSINESS INTERRUPTION) HOWEVER CAUSED AND ON
+# ANY THEORY OF LIABILITY, WHETHER IN CONTRACT, STRICT LIABILITY, OR TORT
+# (INCLUDING NEGLIGENCE OR OTHERWISE) ARISING IN ANY WAY OUT OF THE USE OF THIS
+# SOFTWARE, EVEN IF ADVISED OF THE POSSIBILITY OF SUCH DAMAGE.
+#
+#
+#
 #
 # Script to merge multiple sqlite3 databases that were written using a cSQLiteOutputManager
 #
@@ -14,6 +44,8 @@ import argparse
 import glob
 import sys
 import os
+
+SCHEMAVERSION=2
 
 if __name__ == "__main__":
     parser = argparse.ArgumentParser(description='Merges two or more source databases into destination database. Databases must be in cSQLiteResultManager format.')
@@ -43,12 +75,33 @@ if __name__ == "__main__":
 		# TODO turn this off when everything works! Performance!
 		c.execute("PRAGMA foreign_keys = ON;")
 	
+		c.execute("SELECT value FROM metadata WHERE key='schemaversion';")
+		select_schemaversion = int(c.fetchone()[0])
+		if select_schemaversion != SCHEMAVERSION:
+			print args.databasefiles[0]+' has wrong schema (version '+str(select_schemaversion)+') tool works with schema (version '+str(SCHEMAVERSION)+'). You can try to manually merge your databases.'
+			print 'moving database back to '+args.databasefiles[0]
+			os.rename(args.output,args.databasefiles[0])
+			sys.exit(1)
+
 		for file in args.databasefiles[1:]:
 			sys.stdout.write('Attaching database '+file+'...')
 			# Attach database to merge into source database
 			c.execute("ATTACH DATABASE '"+file+"' AS toMerge;")
 			sys.stdout.write('done\n')
-
+			
+			sys.stdout.write('Checking schema version of '+file+'...')
+			c.execute("SELECT value FROM toMerge.metadata WHERE key='schemaversion';")
+			select_schemaversion = int(c.fetchone()[0])
+			if select_schemaversion != SCHEMAVERSION:
+			    sys.stdout.write('failed\n')
+			    print file+' has wrong schema (version '+str(select_schemaversion)+') tool works with schema (version '+str(SCHEMAVERSION)+'). You can try to manually merge your databases.'
+			    sys.stdout.write('Detaching database '+file+'...')
+			    c.execute("DETACH DATABASE 'toMerge';")
+			    sys.stdout.write('done\n')
+			    continue
+			else:
+			    sys.stdout.write('done\n')
+			    
 			sys.stdout.write('\t deleting duplicate runs (runs that exist in multiple databases) based on the unique runid...')
 			# DELETE DUPLICATES, due to cascading this will delete a lot of data and thus saves time if there are really duplicates
 			c.execute("DELETE FROM toMerge.run WHERE toMerge.run.runid IN (SELECT runid FROM run);")
